Use 64-bit int only when necessary

+++ src/kernels.cu
@@ -13,12 +13,12 @@ __global__ void embedding_kernel(
         fp_t* wte,
         fp_t* wpe,
         fp_t* embeddings,
-        uint64_t batch_size,
-        uint64_t seq_length,
-        uint64_t n_embd) {
+        uint32_t batch_size,
+        uint32_t seq_length,
+        uint32_t n_embd) {
     // Calculate thread ID
-    uint64_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
-    uint64_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
+    uint32_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
+    uint32_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
 
     // Check bounds
     if (idx_batch >= batch_size ||
@@ -30,12 +30,12 @@ __global__ void embedding_kernel(
     id_t token_id = token_ids[idx_batch * seq_length + idx_seq];
 
     // Calculate offsets
-    uint64_t offset_out = (idx_batch * seq_length + idx_seq) * n_embd;
-    uint64_t offset_wte = token_id * n_embd;
-    uint64_t offset_wpe = idx_seq * n_embd;
+    uint64_t offset_out = ((uint64_t)idx_batch * seq_length + idx_seq) * n_embd;
+    uint64_t offset_wte = (uint64_t)token_id * n_embd;
+    uint64_t offset_wpe = (uint64_t)idx_seq * n_embd;
 
     // Perform embedding lookup
-    for (uint64_t i = 0; i < n_embd; i++) {
+    for (uint32_t i = 0; i < n_embd; i++) {
         embeddings[offset_out + i] = wte[offset_wte + i] + wpe[offset_wpe + i];
     }
 }
@@ -45,12 +45,12 @@ __global__ void qkv_projection_kernel(
         fp_t* output,
         fp_t* w_qkv,
         fp_t* b_qkv,
-        uint64_t batch_size,
-        uint64_t seq_length,
-        uint64_t n_embd) {
+        uint32_t batch_size,
+        uint32_t seq_length,
+        uint32_t n_embd) {
     // Calculate thread ID
-    uint64_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
-    uint64_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
+    uint32_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
+    uint32_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
 
     // Check bounds
     if (idx_batch >= batch_size ||
@@ -59,14 +59,14 @@ __global__ void qkv_projection_kernel(
     }
 
     // Get the starting index for the current token
-    uint64_t offset_input = (idx_batch * seq_length + idx_seq) * n_embd;
-    uint64_t offset_output = (idx_batch * seq_length + idx_seq) * (3 * n_embd);
-    uint64_t qkv_size = 3 * n_embd; // Size of Q, K, V for each token
+    uint64_t offset_input  = ((uint64_t) idx_batch * seq_length + idx_seq) * n_embd;
+    uint64_t offset_output = ((uint64_t) idx_batch * seq_length + idx_seq) * (3 * n_embd);
+    uint32_t qkv_size = 3 * n_embd; // Size of Q, K, V for each token
 
     // Perform QKV projection
-    for (uint64_t i = 0; i < qkv_size; i++) {
+    for (uint32_t i = 0; i < qkv_size; i++) {
         fp_t val = b_qkv[i];
-        for (uint64_t j = 0; j < n_embd; j++) {
+        for (uint32_t j = 0; j < n_embd; j++) {
             val += input[offset_input + j] * w_qkv[j * qkv_size + i];
         }
         output[offset_output + i] = val;
@@ -77,12 +77,12 @@ __global__ void layer_normalization_kernel(
         fp_t* input,
         fp_t* gamma,
         fp_t* beta,
-        uint64_t batch_size,
-        uint64_t seq_length,
-        uint64_t n_embd) {
+        uint32_t batch_size,
+        uint32_t seq_length,
+        uint32_t n_embd) {
     // Calculate thread ID
-    uint64_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
-    uint64_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
+    uint32_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
+    uint32_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
 
     // Check bounds
     if (idx_batch >= batch_size ||
@@ -91,18 +91,18 @@ __global__ void layer_normalization_kernel(
     }
 
     // Get the starting index for the current token
-    uint64_t offset_input = (idx_batch * seq_length + idx_seq) * n_embd;
+    uint64_t offset_input = ((uint64_t)idx_batch * seq_length + idx_seq) * n_embd;
 
     // Calculate mean
     fp_t mean = 0.0f;
-    for (uint64_t i = 0; i < n_embd; i++) {
+    for (uint32_t i = 0; i < n_embd; i++) {
         mean += input[offset_input + i];
     }
     mean /= n_embd;
 
     // Calculate variance
     fp_t var = 0.0f;
-    for (uint64_t i = 0; i < n_embd; i++) {
+    for (uint32_t i = 0; i < n_embd; i++) {
         fp_t diff = input[offset_input + i] - mean;
         var += diff * diff;
     }
@@ -112,7 +112,7 @@ __global__ void layer_normalization_kernel(
     const fp_t epsilon = 1e-5f;
     fp_t inv_std = rsqrtf(var + epsilon);
 
-    for (uint64_t i = 0; i < n_embd; i++) {
+    for (uint32_t i = 0; i < n_embd; i++) {
         fp_t normalized = (input[offset_input + i] - mean) * inv_std;
         input[offset_input + i] = gamma[i] * normalized + beta[i];
     }
@@ -121,13 +121,13 @@ __global__ void layer_normalization_kernel(
 __global__ void multi_head_attention_kernel(
         fp_t* qkv,
         fp_t* output,
-        uint64_t batch_size,
-        uint64_t seq_length,
-        uint64_t n_head,
-        uint64_t n_embd) {
+        uint32_t batch_size,
+        uint32_t seq_length,
+        uint32_t n_head,
+        uint32_t n_embd) {
     // Calculate thread ID
-    uint64_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
-    uint64_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
+    uint32_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
+    uint32_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
 
     // Check bounds
     if (idx_batch >= batch_size ||
@@ -136,8 +136,8 @@ __global__ void multi_head_attention_kernel(
     }
 
     // Calculate dimensions
-    uint64_t d_k = n_embd / n_head; // Dimension of each head
-    uint64_t qkv_size = 3 * n_embd; // Size of Q, K, V for each token
+    uint32_t d_k = n_embd / n_head; // Dimension of each head
+    uint32_t qkv_size = 3 * n_embd; // Size of Q, K, V for each token
 
     // Scores register
     fp_t scores[SEQ_LENGTH_MAX];
@@ -145,15 +145,15 @@ __global__ void multi_head_attention_kernel(
     fp_t scale = rsqrtf(d_k);
 
     // Process each attention head
-    for (uint64_t i_head = 0; i_head < n_head; i_head++) {
+    for (uint32_t i_head = 0; i_head < n_head; i_head++) {
         // Calculate attention scores between current token and all other tokens
         fp_t max_val = -INFINITY;
-        for (uint64_t j_token = 0; j_token < seq_length; j_token++) {
+        for (uint32_t j_token = 0; j_token < seq_length; j_token++) {
             // Causal masking: only attend to positions j_token <= i_token
             if (j_token <= idx_seq) {
                 // Compute dot product
                 fp_t dot = 0.0f;
-                for (uint64_t d = 0; d < d_k; d++) {
+                for (uint32_t d = 0; d < d_k; d++) {
                     // Q values for token idx_seq, head i_head
                     // K values for token j_token, head i_head
                     dot += qkv[idx_batch * seq_length * qkv_size + idx_seq * qkv_size + i_head * d_k + d] *
@@ -166,7 +166,7 @@ __global__ void multi_head_attention_kernel(
 
         // Softmax calculation for attention weights
         fp_t sum = 0.0f;
-        for (uint64_t j_token = 0; j_token < seq_length; j_token++) {
+        for (uint32_t j_token = 0; j_token < seq_length; j_token++) {
             // Causal masking: masked tokens have zero weight
             if (j_token <= idx_seq) {
                 scores[j_token] = expf(scores[j_token] - max_val);
@@ -176,14 +176,14 @@ __global__ void multi_head_attention_kernel(
             }
         }
 
-        for (uint64_t j_token = 0; j_token < seq_length; j_token++) {
+        for (uint32_t j_token = 0; j_token < seq_length; j_token++) {
             scores[j_token] /= sum;
         }
 
         // Calculate weighted sum of values
-        for (uint64_t d = 0; d < d_k; d++) {
+        for (uint32_t d = 0; d < d_k; d++) {
             fp_t weighted_sum = 0.0f;
-            for (uint64_t j_token = 0; j_token < seq_length; j_token++) {
+            for (uint32_t j_token = 0; j_token < seq_length; j_token++) {
                 // Get V values for token j_token, head i_head
                 weighted_sum += scores[j_token] *
                                 qkv[idx_batch * seq_length * qkv_size + j_token * qkv_size + 2 * n_embd + i_head * d_k + d];
@@ -199,12 +199,12 @@ __global__ void final_projection_kernel(
         fp_t* output,
         fp_t* w_proj,
         fp_t* b_proj,
-        uint64_t batch_size,
-        uint64_t seq_length,
-        uint64_t n_embd) {
+        uint32_t batch_size,
+        uint32_t seq_length,
+        uint32_t n_embd) {
     // Calculate thread ID
-    uint64_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
-    uint64_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
+    uint32_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
+    uint32_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
 
     // Check bounds
     if (idx_batch >= batch_size ||
@@ -213,12 +213,12 @@ __global__ void final_projection_kernel(
     }
 
     // Get the starting index for the current token
-    uint64_t offset_input = (idx_batch * seq_length + idx_seq) * n_embd;
+    uint64_t offset_input = ((uint64_t)idx_batch * seq_length + idx_seq) * n_embd;
 
     // Perform final projection
-    for (uint64_t i = 0; i < n_embd; i++) {
+    for (uint32_t i = 0; i < n_embd; i++) {
         fp_t val = b_proj[i];
-        for (uint64_t j = 0; j < n_embd; j++) {
+        for (uint32_t j = 0; j < n_embd; j++) {
             val += input[offset_input + j] * w_proj[j * n_embd + i];
         }
         output[offset_input + i] = val;
@@ -229,12 +229,12 @@ __global__ void add_residual_kernel(
         fp_t* input,
         fp_t* residual,
         fp_t* output,
-        uint64_t batch_size,
-        uint64_t seq_length,
-        uint64_t n_embd) {
+        uint32_t batch_size,
+        uint32_t seq_length,
+        uint32_t n_embd) {
     // Calculate thread ID
-    uint64_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
-    uint64_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
+    uint32_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
+    uint32_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
 
     // Check bounds
     if (idx_batch >= batch_size ||
@@ -243,10 +243,10 @@ __global__ void add_residual_kernel(
     }
 
     // Get the starting index for the current token
-    uint64_t offset = (idx_batch * seq_length + idx_seq) * n_embd;
+    uint64_t offset = ((uint64_t)idx_batch * seq_length + idx_seq) * n_embd;
 
     // Add residual connection
-    for (uint64_t i = 0; i < n_embd; i++) {
+    for (uint32_t i = 0; i < n_embd; i++) {
         output[offset + i] = input[offset + i] + residual[offset + i];
     }
 }
@@ -258,12 +258,12 @@ __global__ void mlp_kernel(
         fp_t* b_fc, 
         fp_t* w_proj,
         fp_t* b_proj,
-        uint64_t batch_size,
-        uint64_t seq_length,
-        uint64_t n_embd) {
+        uint32_t batch_size,
+        uint32_t seq_length,
+        uint32_t n_embd) {
     // Calculate thread ID
-    uint64_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
-    uint64_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
+    uint32_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
+    uint32_t idx_seq   = blockIdx.y * blockDim.y + threadIdx.y;
 
     // Check bounds
     if (idx_batch >= batch_size ||
@@ -272,25 +272,25 @@ __global__ void mlp_kernel(
     }
 
     // Intermediate register
-    uint64_t intermediate_size = 4 * n_embd;
+    uint32_t intermediate_size = 4 * n_embd;
     fp_t intermediate[INTERMEDIATE_SIZE_MAX];
 
     // Get the starting index for the current token
-    uint64_t offset_input = (idx_batch * seq_length + idx_seq) * n_embd;
+    uint64_t offset_input = ((uint64_t)idx_batch * seq_length + idx_seq) * n_embd;
 
     // Compute feedforward layer
-    for (uint64_t i = 0; i < intermediate_size; i++) {
+    for (uint32_t i = 0; i < intermediate_size; i++) {
         fp_t val = b_fc[i];
-        for (uint64_t j = 0; j < n_embd; j++) {
+        for (uint32_t j = 0; j < n_embd; j++) {
             val += input[offset_input + j] * w_fc[j * intermediate_size + i];
         }
         intermediate[i] = gelu(val);
     }
 
     // Compute projection back to hidden size
-    for (uint64_t i = 0; i < n_embd; i++) {
+    for (uint32_t i = 0; i < n_embd; i++) {
         fp_t val = b_proj[i];
-        for (uint64_t j = 0; j < intermediate_size; j++) {
+        for (uint32_t j = 0; j < intermediate_size; j++) {
             val += intermediate[j] * w_proj[j * n_embd + i];
         }
         output[offset_input + i] = val;
@@ -302,13 +302,13 @@ __global__ void lm_head_kernel(
         fp_t* logits,
         fp_t* weights,
         fp_t* biases,
-        uint64_t batch_size,
-        uint64_t seq_length,
-        uint64_t n_vocab,
-        uint64_t n_embd) {
+        uint32_t batch_size,
+        uint32_t seq_length,
+        uint32_t n_vocab,
+        uint32_t n_embd) {
     // Calculate thread ID
-    uint64_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
-    uint64_t idx_vocab = blockIdx.y * blockDim.y + threadIdx.y;
+    uint32_t idx_batch = blockIdx.x * blockDim.x + threadIdx.x;
+    uint32_t idx_vocab = blockIdx.y * blockDim.y + threadIdx.y;
 
     // Check bounds
     if (idx_batch >= batch_size ||
@@ -317,15 +317,15 @@ __global__ void lm_head_kernel(
     }
 
     // Calculate output index
-    uint64_t idx_out = (idx_batch * n_vocab + idx_vocab);
+    uint64_t idx_out = ((uint64_t)idx_batch * n_vocab + idx_vocab);
 
     // Get the starting index for the current token
-    uint64_t offset_input =  (idx_batch * seq_length + (seq_length - 1)) * n_embd;
-    uint64_t offset_weights = idx_vocab * n_embd;
+    uint64_t offset_input =  ((uint64_t)idx_batch * seq_length + (seq_length - 1)) * n_embd;
+    uint64_t offset_weights = (uint64_t)idx_vocab * n_embd;
 
     // Compute logits
     logits[idx_out] = biases ? biases[idx_vocab] : 0.0f;
-    for (uint64_t i = 0; i < n_embd; i++) {
+    for (uint32_t i = 0; i < n_embd; i++) {
         logits[idx_out] += hidden_state[offset_input + i] * weights[offset_weights + i];
     }
 }
